{"types":"C","runs":13}

+++ Problem_16.java
@@ -7,7 +7,7 @@ public class Problem_16 extends ConsoleProgram {
 		int n1 = readInt ("enter n1: ");
 		int n2 = readInt ("enter n2: ");
 		int n3 = readInt ("enter n3: ");
-		double harmonic = (double) 1 / (1 / n1 + 1 / n2 + 1 / n3);
+		double harmonic = (int) 1 / (1 / n1 + 1 / n2 + 1 / n3);
 		println("three integers harmonic: "+ harmonic + ".");
 	}
 }
